ML MODEL Complete
Completed Designing and Coding Machine Learning Model to detect Impacts, currently at a 85.71% accuracy. Will be doing more testing and data collection in order to improve accuracy of results. Will be working on classifying impacts as concussions or not using thresholds for project presentation. Need to update Hardware design for better power usage.